+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 1.8.9 (2025-05-19)
+
+- Fix deprecation warning in Click 8.2. [[#215](https://github.com/ewels/rich-click/pull/215)] ([@finsberg](https://github.com/finsberg))
+- Fix typing incompatibilities with Click 8.2.
+- Added Click 8.2's support for `Parameter.deprecated: str | bool`
+
 ## Version 1.8.8 (2025-03-09)
 
 - Make text wrap instead of using ellipses for overflowing metavars in options tables.